Spellbar updates, bug fix ```diff + Fix for using Esc as a macro hotkey + Added configurable hotkey support for spellbar + Updated spellbar assistant settings ```
