+++ src/ClassicUO.Client/Game/Scenes/GameSceneInputHandler.cs
@@ -1328,6 +1328,8 @@ namespace ClassicUO.Game.Scenes
             {
                 return;
             }
+            
+            SpellBarManager.KeyPress(e.keysym.sym, e.keysym.mod);
 
             if (e.keysym.sym == SDL.SDL_Keycode.SDLK_ESCAPE && TargetManager.IsTargeting)
             {
@@ -1702,6 +1704,8 @@ namespace ClassicUO.Game.Scenes
         internal override void OnControllerButtonDown(SDL.SDL_ControllerButtonEvent e)
         {
             base.OnControllerButtonDown(e);
+            
+            SpellBarManager.ControllerInput((SDL.SDL_GameControllerButton)e.button);
 
             if (World.InGame && (UIManager.KeyboardFocusControl == UIManager.SystemChat.TextBoxControl || UIManager.KeyboardFocusControl == null))
             {
