+++ src/ClassicUO.Client/Game/UI/Gumps/AssistantGump.cs
@@ -134,7 +134,20 @@ public class AssistantGump : BaseOptionsGump
         PositionHelper.Reset();
         
         scroll.Add(PositionHelper.PositionControl(new HttpClickableLink("SpellBar Wiki", "https://github.com/bittiez/TazUO/wiki/TazUO.SpellBar", ThemeSettings.TEXT_FONT_COLOR)));
-        scroll.Add(PositionHelper.PositionControl(TextBox.GetOne(".", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, ThemeSettings.TEXT_FONT_COLOR, TextBox.RTLOptions.Default(MainContent.RightWidth - 20))));
+        PositionHelper.BlankLine();
+        
+        scroll.Add(PositionHelper.PositionControl(new CheckboxWithLabel("Enable spellbar", 0, SpellBarManager.IsEnabled(), (b) =>
+        {
+            if (SpellBarManager.ToggleEnabled())
+            {
+                UIManager.Add(new SpellBar.SpellBar());
+            }
+            else
+            {
+                SpellBar.SpellBar.Instance?.Dispose();
+            }
+            
+        })));
         PositionHelper.BlankLine();
         
         ModernButton b;
@@ -146,7 +159,7 @@ public class AssistantGump : BaseOptionsGump
         };
 
         ModernButton bb;
-        scroll.Add(PositionHelper.ToRightOf(bb = new ModernButton(0, 0, 100, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "Remove row", ThemeSettings.BUTTON_FONT_COLOR), b));
+        scroll.Add(PositionHelper.ToRightOf(bb = new ModernButton(0, 0, 150, ThemeSettings.CHECKBOX_SIZE, ButtonAction.Default, "Remove row", ThemeSettings.BUTTON_FONT_COLOR), b));
         bb.SetTooltip("This will remove the last row. If you have 5 rows, row 5 will be removed.");
         bb.MouseUp += (s, e) =>
         {
@@ -154,6 +167,34 @@ public class AssistantGump : BaseOptionsGump
                 SpellBarManager.SpellBarRows.RemoveAt(SpellBarManager.SpellBarRows.Count - 1);
             SpellBar.SpellBar.Instance?.Build();
         };
+
+        var controllerHotkeys = SpellBarManager.GetControllerButtons();
+        var hotkeys = SpellBarManager.GetHotKeys();
+        var keymods = SpellBarManager.GetModKeys();
+        
+        
+        for(var c = 0; c < 10; c++)
+        {
+            PositionHelper.BlankLine();
+            Control tb;
+            scroll.Add(tb = PositionHelper.PositionControl(TextBox.GetOne($"Row {c} hotkeys: ", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, ThemeSettings.TEXT_FONT_COLOR, TextBox.RTLOptions.Default())));
+            
+            HotkeyBox hotkey = new();
+            var c1 = c;
+
+            hotkey.HotkeyChanged += (s, e) =>
+            {
+                SpellBarManager.SetButtons(c1, hotkey.Mod, hotkey.Key, hotkey.Buttons);
+            };
+
+            if (controllerHotkeys.Length > c)
+                hotkey.SetButtons(controllerHotkeys[c]);
+            
+            if(hotkeys.Length > c && keymods.Length > c)
+                hotkey.SetKey(hotkeys[c], keymods[c]);
+
+            scroll.Add(PositionHelper.ToRightOf(hotkey, tb));
+        }
     }
     
     private enum PAGE
