+++ src/ClassicUO.Client/Game/UI/Gumps/BaseOptionsGump.cs
@@ -535,6 +535,258 @@ public class BaseOptionsGump : Gump
             }
         }
     }
+    
+    protected class HotkeyBox : Control
+    {
+        private bool _actived;
+        private readonly ModernButton _buttonOK, _buttonCancel;
+        private readonly TextBox _label;
+
+        public HotkeyBox()
+        {
+            CanMove = false;
+            AcceptMouseInput = true;
+            AcceptKeyboardInput = true;
+
+            Width = 300;
+            Height = 40;
+
+            AlphaBlendControl bg = new AlphaBlendControl()
+            {
+                Width = 150,
+                Height = 40,
+                AcceptMouseInput = true
+            };
+
+            Add(bg);
+            bg.MouseUp += LabelOnMouseUp;
+
+            Add(_label = TextBox.GetOne("None", ThemeSettings.FONT, ThemeSettings.STANDARD_TEXT_SIZE, ThemeSettings.TEXT_FONT_COLOR, TextBox.RTLOptions.DefaultCentered(150)));
+            _label.Y = (bg.Height >> 1) - (_label.Height >> 1);
+
+            _label.MouseUp += LabelOnMouseUp;
+
+            Add
+            (
+                _buttonOK = new ModernButton(152, 0, 75, 40, ButtonAction.Activate, "Save", ThemeSettings.BUTTON_FONT_COLOR)
+                {
+                    ButtonParameter = (int)ButtonState.Ok
+                }
+            );
+
+            Add
+            (
+                _buttonCancel = new ModernButton(_buttonOK.Bounds.Right + 5, 0, 75, 40, ButtonAction.Activate, "Cancel", ThemeSettings.BUTTON_FONT_COLOR)
+                {
+                    ButtonParameter = (int)ButtonState.Cancel
+                }
+            );
+
+            WantUpdateSize = false;
+            IsActive = false;
+        }
+
+        public SDL.SDL_Keycode Key { get; private set; }
+        public SDL.SDL_GameControllerButton[] Buttons { get; private set; }
+        public MouseButtonType MouseButton { get; private set; }
+        public bool WheelScroll { get; private set; }
+        public bool WheelUp { get; private set; }
+        public SDL.SDL_Keymod Mod { get; private set; }
+
+        public bool IsActive
+        {
+            get => _actived;
+            set
+            {
+                _actived = value;
+
+                if (value)
+                {
+                    _buttonOK.IsVisible = _buttonCancel.IsVisible = true;
+                    _buttonOK.IsEnabled = _buttonCancel.IsEnabled = true;
+                }
+                else
+                {
+                    _buttonOK.IsVisible = _buttonCancel.IsVisible = false;
+                    _buttonOK.IsEnabled = _buttonCancel.IsEnabled = false;
+                }
+            }
+        }
+
+        public event EventHandler HotkeyChanged, HotkeyCancelled;
+
+        protected override void OnControllerButtonDown(SDL.SDL_GameControllerButton button)
+        {
+            if (IsActive)
+            {
+                SetButtons(Controller.PressedButtons());
+            }
+        }
+
+        protected override void OnKeyDown(SDL.SDL_Keycode key, SDL.SDL_Keymod mod)
+        {
+            if (IsActive)
+            {
+                SetKey(key, mod);
+            }
+        }
+
+        public void SetButtons(SDL.SDL_GameControllerButton[] buttons)
+        {
+            ResetBinding();
+            Buttons = buttons;
+            _label.Text = Controller.GetButtonNames(buttons);
+        }
+
+        public void SetKey(SDL.SDL_Keycode key, SDL.SDL_Keymod mod)
+        {
+            if (key == SDL.SDL_Keycode.SDLK_UNKNOWN && mod == SDL.SDL_Keymod.KMOD_NONE)
+            {
+                ResetBinding();
+
+                Key = key;
+                Mod = mod;
+            }
+            else
+            {
+                string newvalue = KeysTranslator.TryGetKey(key, mod);
+
+                if (!string.IsNullOrEmpty(newvalue) && key != SDL.SDL_Keycode.SDLK_UNKNOWN)
+                {
+                    ResetBinding();
+
+                    Key = key;
+                    Mod = mod;
+                    _label.Text = newvalue;
+                }
+            }
+        }
+
+        protected override void OnMouseDown(int x, int y, MouseButtonType button)
+        {
+            if (button == MouseButtonType.Middle || button == MouseButtonType.XButton1 || button == MouseButtonType.XButton2)
+            {
+                SDL.SDL_Keymod mod = SDL.SDL_Keymod.KMOD_NONE;
+
+                if (Keyboard.Alt)
+                {
+                    mod |= SDL.SDL_Keymod.KMOD_ALT;
+                }
+
+                if (Keyboard.Shift)
+                {
+                    mod |= SDL.SDL_Keymod.KMOD_SHIFT;
+                }
+
+                if (Keyboard.Ctrl)
+                {
+                    mod |= SDL.SDL_Keymod.KMOD_CTRL;
+                }
+
+                SetMouseButton(button, mod);
+            }
+        }
+
+        public void SetMouseButton(MouseButtonType button, SDL.SDL_Keymod mod)
+        {
+            string newvalue = KeysTranslator.GetMouseButton(button, mod);
+
+            if (!string.IsNullOrEmpty(newvalue) && button != MouseButtonType.None)
+            {
+                ResetBinding();
+
+                MouseButton = button;
+                Mod = mod;
+                _label.Text = newvalue;
+            }
+        }
+
+        protected override void OnMouseWheel(MouseEventType delta)
+        {
+            SDL.SDL_Keymod mod = SDL.SDL_Keymod.KMOD_NONE;
+
+            if (Keyboard.Alt)
+            {
+                mod |= SDL.SDL_Keymod.KMOD_ALT;
+            }
+
+            if (Keyboard.Shift)
+            {
+                mod |= SDL.SDL_Keymod.KMOD_SHIFT;
+            }
+
+            if (Keyboard.Ctrl)
+            {
+                mod |= SDL.SDL_Keymod.KMOD_CTRL;
+            }
+
+            if (delta == MouseEventType.WheelScrollUp)
+            {
+                SetMouseWheel(true, mod);
+            }
+            else if (delta == MouseEventType.WheelScrollDown)
+            {
+                SetMouseWheel(false, mod);
+            }
+        }
+
+        public void SetMouseWheel(bool wheelUp, SDL.SDL_Keymod mod)
+        {
+            string newvalue = KeysTranslator.GetMouseWheel(wheelUp, mod);
+
+            if (!string.IsNullOrEmpty(newvalue))
+            {
+                ResetBinding();
+
+                WheelScroll = true;
+                WheelUp = wheelUp;
+                Mod = mod;
+                _label.Text = newvalue;
+            }
+        }
+
+        private void ResetBinding()
+        {
+            Key = 0;
+            MouseButton = MouseButtonType.None;
+            WheelScroll = false;
+            Mod = 0;
+            _label.Text = "None";
+            Buttons = null;
+        }
+
+        private void LabelOnMouseUp(object sender, MouseEventArgs e)
+        {
+            IsActive = true;
+            SetKeyboardFocus();
+        }
+
+        public override void OnButtonClick(int buttonID)
+        {
+            switch ((ButtonState)buttonID)
+            {
+                case ButtonState.Ok: HotkeyChanged.Raise(this); break;
+
+                case ButtonState.Cancel:
+                    _label.Text = "None";
+
+                    HotkeyCancelled.Raise(this);
+
+                    Key = SDL.SDL_Keycode.SDLK_UNKNOWN;
+                    Mod = SDL.SDL_Keymod.KMOD_NONE;
+
+                    break;
+            }
+
+            IsActive = false;
+        }
+
+        private enum ButtonState
+        {
+            Ok,
+            Cancel
+        }
+    }
 
     public class InputField : Control
     {
